requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
tqdm>=4.65.0
pillow>=10.0.0
psutil>=5.9.0
qwen-vl-utils>=0.0.1
accelerate>=0.20.0
safetensors>=0.3.0 